feature(controller):rename userController to authController
